+++ src/app/components/forms/add/index.tsx
@@ -1,80 +1,72 @@
-import { Modal } from "@/app/components/modal";
-import { Select } from "@/app/components/select";
-import { updateTodo } from "@/app/operations";
+"use client";
+import { Select } from "@/app/components/forms/select";
+import { CustomTooltip } from "@/app/components/forms/tooltip";
+import { Modal } from "@/app/components/surfaces/modal";
 import { Category, Importance, Priority, Todo } from "@/app/types";
-import { CustomTooltip } from "@/app/utils/custom-tooltip";
 import { formatISO, parseISO } from "date-fns";
 import React, { useState } from "react";
 import styles from "./styles.module.css";
 
 type Props = {
-  todo: Todo;
   onClose: () => void;
-  onUpdateTodo: (updatedTodo: Todo) => void;
+  onAddTodo: (todo: Omit<Todo, "id" | "createdAt">) => void;
   categories: Category[];
   priorities: Priority[];
   importances: Importance[];
 };
 
-export const EditTodoModal = ({
-  todo,
+export const AddTodoModal = ({
   onClose,
-  onUpdateTodo,
+  onAddTodo,
   categories,
   priorities,
   importances,
 }: Props) => {
-  const [description, setDescription] = useState<string>(
-    todo.description || ""
-  );
-  const [selectedCategory, setSelectedCategory] = useState<string>(
-    todo.categoryId || ""
-  );
-  const [selectedPriority, setSelectedPriority] = useState<string>(
-    todo.priorityId || ""
-  );
-  const [selectedImportance, setSelectedImportance] = useState<string>(
-    todo.importanceId || ""
-  );
-  const [deadline, setDeadline] = useState<string>(
-    todo.deadline
-      ? formatISO(parseISO(todo.deadline), { representation: "date" })
-      : ""
-  );
+  const [description, setDescription] = useState("");
+  const [selectedCategory, setSelectedCategory] = useState("");
+  const [selectedPriority, setSelectedPriority] = useState("");
+  const [selectedImportance, setSelectedImportance] = useState("");
+  const [deadline, setDeadline] = useState("");
 
-  const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
+  const handleSubmit = (e: React.FormEvent) => {
     e.preventDefault();
-    const updatedTodo: Todo = {
-      ...todo,
+    const todo: Omit<Todo, "id" | "createdAt"> = {
       description,
       categoryId: selectedCategory,
       priorityId: selectedPriority,
       importanceId: selectedImportance,
       deadline: deadline ? formatISO(parseISO(deadline)) : null,
+      status: {
+        key: "incomplete",
+        name: "未完了",
+        id: "cluruuvds000cxhqzcyrbabnm",
+      },
     };
-    try {
-      const response = await updateTodo(updatedTodo);
-      onUpdateTodo(response);
-      onClose();
-    } catch (error) {
-      alert(`保存処理中にエラーが発生しました: ${error}`);
-    }
+    onAddTodo(todo);
+    resetForm();
+    onClose();
+  };
+
+  const resetForm = () => {
+    setDescription("");
+    setSelectedCategory("");
+    setSelectedPriority("");
+    setSelectedImportance("");
+    setDeadline("");
   };
 
   return (
     <Modal>
-      <form onSubmit={handleSubmit} className={styles.formContainer}>
-        <div className={styles.inputRow}>
-          <input
-            type="text"
-            value={description}
-            onChange={(e) => setDescription(e.target.value)}
-            placeholder="Todoを入力してください"
-            required
-            className={styles.input}
-          />
-        </div>
-        <div className={styles.inputRow}>
+      <form onSubmit={handleSubmit} className={styles.form}>
+        <input
+          type="text"
+          value={description}
+          onChange={(e) => setDescription(e.target.value)}
+          placeholder="Todoを入力してください"
+          required
+          className={styles.inputText}
+        />
+        <div className={styles.inputContainer}>
           <Select
             options={categories}
             value={selectedCategory}
@@ -96,10 +88,11 @@ export const EditTodoModal = ({
           <CustomTooltip text="期限を選択してください">
             <input
               type="date"
+              id="deadline"
               value={deadline}
               onChange={(e) => setDeadline(e.target.value)}
               required
-              className={styles.input}
+              className={styles.inputDeadline}
             />
           </CustomTooltip>
         </div>
@@ -111,8 +104,8 @@ export const EditTodoModal = ({
           >
             キャンセル
           </button>
-          <button type="submit" className={styles.editButton}>
-            保存
+          <button className={styles.submitButton} type="submit">
+            追加
           </button>
         </div>
       </form>
