+++ src/app/components/forms/edit/index.tsx
@@ -1,72 +1,80 @@
-"use client";
-import { Modal } from "@/app/components/modal";
-import { Select } from "@/app/components/select";
+import { Select } from "@/app/components/forms/select";
+import { CustomTooltip } from "@/app/components/forms/tooltip";
+import { Modal } from "@/app/components/surfaces/modal";
+import { updateTodo } from "@/app/operations";
 import { Category, Importance, Priority, Todo } from "@/app/types";
-import { CustomTooltip } from "@/app/utils/custom-tooltip";
 import { formatISO, parseISO } from "date-fns";
 import React, { useState } from "react";
 import styles from "./styles.module.css";
 
 type Props = {
+  todo: Todo;
   onClose: () => void;
-  onAddTodo: (todo: Omit<Todo, "id" | "createdAt">) => void;
+  onUpdateTodo: (updatedTodo: Todo) => void;
   categories: Category[];
   priorities: Priority[];
   importances: Importance[];
 };
 
-export const AddTodoModal = ({
+export const EditTodoModal = ({
+  todo,
   onClose,
-  onAddTodo,
+  onUpdateTodo,
   categories,
   priorities,
   importances,
 }: Props) => {
-  const [description, setDescription] = useState("");
-  const [selectedCategory, setSelectedCategory] = useState("");
-  const [selectedPriority, setSelectedPriority] = useState("");
-  const [selectedImportance, setSelectedImportance] = useState("");
-  const [deadline, setDeadline] = useState("");
+  const [description, setDescription] = useState<string>(
+    todo.description || ""
+  );
+  const [selectedCategory, setSelectedCategory] = useState<string>(
+    todo.categoryId || ""
+  );
+  const [selectedPriority, setSelectedPriority] = useState<string>(
+    todo.priorityId || ""
+  );
+  const [selectedImportance, setSelectedImportance] = useState<string>(
+    todo.importanceId || ""
+  );
+  const [deadline, setDeadline] = useState<string>(
+    todo.deadline
+      ? formatISO(parseISO(todo.deadline), { representation: "date" })
+      : ""
+  );
 
-  const handleSubmit = (e: React.FormEvent) => {
+  const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
-    const todo: Omit<Todo, "id" | "createdAt"> = {
+    const updatedTodo: Todo = {
+      ...todo,
       description,
       categoryId: selectedCategory,
       priorityId: selectedPriority,
       importanceId: selectedImportance,
       deadline: deadline ? formatISO(parseISO(deadline)) : null,
-      status: {
-        key: "incomplete",
-        name: "未完了",
-        id: "cluruuvds000cxhqzcyrbabnm",
-      },
     };
-    onAddTodo(todo);
-    resetForm();
-    onClose();
-  };
-
-  const resetForm = () => {
-    setDescription("");
-    setSelectedCategory("");
-    setSelectedPriority("");
-    setSelectedImportance("");
-    setDeadline("");
+    try {
+      const response = await updateTodo(updatedTodo);
+      onUpdateTodo(response);
+      onClose();
+    } catch (error) {
+      alert(`保存処理中にエラーが発生しました: ${error}`);
+    }
   };
 
   return (
     <Modal>
-      <form onSubmit={handleSubmit} className={styles.form}>
-        <input
-          type="text"
-          value={description}
-          onChange={(e) => setDescription(e.target.value)}
-          placeholder="Todoを入力してください"
-          required
-          className={styles.inputText}
-        />
-        <div className={styles.inputContainer}>
+      <form onSubmit={handleSubmit} className={styles.formContainer}>
+        <div className={styles.inputRow}>
+          <input
+            type="text"
+            value={description}
+            onChange={(e) => setDescription(e.target.value)}
+            placeholder="Todoを入力してください"
+            required
+            className={styles.input}
+          />
+        </div>
+        <div className={styles.inputRow}>
           <Select
             options={categories}
             value={selectedCategory}
@@ -88,11 +96,10 @@ export const AddTodoModal = ({
           <CustomTooltip text="期限を選択してください">
             <input
               type="date"
-              id="deadline"
               value={deadline}
               onChange={(e) => setDeadline(e.target.value)}
               required
-              className={styles.inputDeadline}
+              className={styles.input}
             />
           </CustomTooltip>
         </div>
@@ -104,8 +111,8 @@ export const AddTodoModal = ({
           >
             キャンセル
           </button>
-          <button className={styles.submitButton} type="submit">
-            追加
+          <button type="submit" className={styles.editButton}>
+            保存
           </button>
         </div>
       </form>
