+++ src/app/components/main/status-tabs/styles.module.css
